+++ lib/jei/document.rb
@@ -11,6 +11,8 @@ module Jei
     #
     # @param [Object] resource
     # @param [Hash<Symbol, Object>] options
+    # @option options [Array<Hash<Symbol, Object>>] :errors override primary
+    #   data with an array of error objects
     # @option options [Hash<String, String>] :fields restrict resource
     #   attributes and relationships to a user-defined set of fields
     # @option options [String] :include a list of relationship paths